+++ src/main/java/fxlauncher/CreateManifest.java
@@ -1,10 +1,12 @@
 package fxlauncher;
 
 import com.sun.javafx.application.ParametersImpl;
+import com.sun.javafx.fxml.builder.URLBuilder;
 
 import javax.xml.bind.JAXB;
 import java.io.IOException;
 import java.net.URI;
+import java.net.URISyntaxException;
 import java.nio.file.*;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
@@ -20,7 +22,7 @@ public class CreateManifest {
         includeExtensions.addAll(Arrays.asList("jar", "war"));
     }
 
-    public static void main(String[] args) throws IOException {
+    public static void main(String[] args) throws IOException, URISyntaxException {
         URI baseURI = URI.create(args[0]);
         String launchClass = args[1];
         Path appPath = Paths.get(args[2]);
@@ -28,6 +30,9 @@ public class CreateManifest {
         String cacheDir = null;
         Boolean acceptDowngrade = null;
         String parameters = null;
+        String whatsNew = null;
+        String preloadNativeLibraries = null;
+        Boolean lingeringUpdateScreen = false;
 
         if (args.length > 3) {
             // Parse named parameters
@@ -45,6 +50,18 @@ public class CreateManifest {
                 if (named.containsKey("accept-downgrade"))
                     acceptDowngrade = Boolean.valueOf(named.get("accept-downgrade"));
 
+                // Configure preload native libraries
+                if (named.containsKey("preload-native-libraries"))
+                    preloadNativeLibraries = named.get("preload-native-libraries");
+
+                // Should the update screen stay until the primary stage is shown?
+                if (named.containsKey("lingering-update-screen"))
+                    lingeringUpdateScreen = Boolean.valueOf(named.get("lingering-update-screen"));
+
+                // Configure the whats-new option
+                if (named.containsKey("whats-new"))
+                    whatsNew = named.get("whats-new");
+
                 // Add additional files with these extensions to manifest
                 if (named.containsKey("include-extensions"))
                     includeExtensions.addAll(
@@ -60,6 +77,9 @@ public class CreateManifest {
                 if (raw.startsWith("--cache-dir=")) continue;
                 if (raw.startsWith("--accept-downgrade=")) continue;
                 if (raw.startsWith("--include-extensions=")) continue;
+                if (raw.startsWith("--preload-native-libraries=")) continue;
+                if (raw.startsWith("--whats-new")) continue;
+                if (raw.startsWith("--lingering-update-screen")) continue;
                 if (rest.length() > 0) rest.append(" ");
                 rest.append(raw);
             }
@@ -73,16 +93,21 @@ public class CreateManifest {
         if (cacheDir != null) manifest.cacheDir = cacheDir;
         if (acceptDowngrade != null) manifest.acceptDowngrade = acceptDowngrade;
         if (parameters != null) manifest.parameters = parameters;
-
+        if (preloadNativeLibraries != null) manifest.preloadNativeLibraries = preloadNativeLibraries;
+        if (whatsNew != null) manifest.whatsNewPage = whatsNew;
+        manifest.lingeringUpdateScreen = lingeringUpdateScreen;
         JAXB.marshal(manifest, appPath.resolve("app.xml").toFile());
     }
 
-    public static FXManifest create(URI baseURI, String launchClass, Path appPath) throws IOException {
+    public static FXManifest create(URI baseURI, String launchClass, Path appPath) throws IOException, URISyntaxException {
         FXManifest manifest = new FXManifest();
         manifest.ts = System.currentTimeMillis();
         manifest.uri = baseURI;
         manifest.launchClass = launchClass;
 
+        if(!manifest.uri.getPath().endsWith("/")) {
+            manifest.uri = new URI(String.format("%s/", baseURI.toString()));
+        }
         Files.walkFileTree(appPath, new SimpleFileVisitor<Path>() {
             public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
                 if (!Files.isDirectory(file) && shouldIncludeInManifest(file) && !file.getFileName().toString().startsWith("fxlauncher"))
@@ -92,6 +117,16 @@ public class CreateManifest {
         });
 
         return manifest;
+    }
+
+    /**
+     * Add the includeExtensions to the default list of "war" and "jar".
+     *
+     * Allthough the method is called setIncludeExtensions, it actually does an addAll.
+     * @param includeExtensions
+     */
+    public static void setIncludeExtensions(List<String> includeExtensions) {
+        CreateManifest.includeExtensions.addAll(includeExtensions);
     }
 
     private static boolean shouldIncludeInManifest(Path file) {